# demo/todo-nest/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "登录",
        "tags": [
          "鉴权"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "注册",
        "tags": [
          "鉴权"
        ]
      }
    },
    "/todos": {
      "get": {
        "operationId": "TodoController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取 Todo 列表",
        "tags": [
          "Todo"
        ]
      },
      "post": {
        "operationId": "TodoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建 Todo",
        "tags": [
          "Todo"
        ]
      }
    },
    "/todos/with-user": {
      "get": {
        "operationId": "TodoController_listWithUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "联表查询（INNER JOIN users）",
        "tags": [
          "Todo"
        ]
      }
    },
    "/todos/page": {
      "get": {
        "operationId": "TodoController_listPaged",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "completed",
            "required": false,
            "in": "query",
            "description": "按完成状态过滤",
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "分页查询（含 Redis 缓存）",
        "tags": [
          "Todo"
        ]
      }
    },
    "/todos/stats/completed": {
      "get": {
        "operationId": "TodoController_statsCompleted",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "按完成状态分组统计",
        "tags": [
          "Todo"
        ]
      }
    },
    "/todos/tx/toggle": {
      "post": {
        "operationId": "TodoController_txToggle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TxToggleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "事务批量切换完成状态",
        "tags": [
          "Todo"
        ]
      }
    },
    "/todos/{id}": {
      "get": {
        "operationId": "TodoController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "按 ID 获取",
        "tags": [
          "Todo"
        ]
      },
      "put": {
        "operationId": "TodoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新 Todo",
        "tags": [
          "Todo"
        ]
      },
      "delete": {
        "operationId": "TodoController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除 Todo",
        "tags": [
          "Todo"
        ]
      }
    }
  },
  "info": {
    "title": "Todo API",
    "description": "基于 NestJS + Prisma + Redis 的 Todo 接口",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "newuser"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        }
      },
      "TxToggleDto": {
        "type": "object",
        "properties": {
          "ids": {
            "example": [
              1,
              2,
              3
            ],
            "description": "要批量切换完成状态的 Todo ID 列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateTodoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "学习 NestJS"
          },
          "userId": {
            "type": "number",
            "description": "所属用户 ID，不传时使用当前登录用户"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateTodoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "completed": {
            "type": "boolean"
          }
        }
      }
    }
  }
}